+++ nuxt.config.js
@@ -32,7 +32,7 @@ export default {
 
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
   plugins: [
-    '~/plugins/helper.js'
+    '~/plugins/helper.ts'
   ],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
